add button logout on header
