+++ nuse-frontend/src/components/layout/index.js
@@ -1,13 +1,7 @@
-import { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { useRouter } from 'next/router';
-
-import { useDispatch } from 'react-redux';
-import { logout } from '../../../redux/actions';
 
 import { Container } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import Cookies from 'js-cookie';
 
 import Footer from './Footer';
 import Header from './Header';
@@ -21,15 +15,6 @@ const useStyles = makeStyles((theme) => ({
 
 const Layout = ({ t, children }) => {
   const classes = useStyles();
-  const router = useRouter();
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    if (!Cookies.get('token')) {
-      dispatch(logout());
-      router.push('/login');
-    }
-  }, [Cookies.get('token')]);
 
   return (
     <>
